+++ instrumentation/runtime-metrics/runtime-metrics-java8/library/src/main/java/io/opentelemetry/instrumentation/runtimemetrics/java8/internal/JmxRuntimeMetricsUtil.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package io.opentelemetry.instrumentation.runtimemetrics.java8;
+package io.opentelemetry.instrumentation.runtimemetrics.java8.internal;
 
 import io.opentelemetry.api.OpenTelemetry;
 import io.opentelemetry.api.metrics.Meter;
@@ -12,6 +12,10 @@ import io.opentelemetry.instrumentation.api.internal.EmbeddedInstrumentationProp
 import java.util.List;
 import javax.annotation.Nullable;
 
+/**
+ * This class is internal and is hence not for public use. Its APIs are unstable and can change at
+ * any time.
+ */
 public class JmxRuntimeMetricsUtil {
 
   private static final String INSTRUMENTATION_NAME = "io.opentelemetry.runtime-metrics-java8";
@@ -20,7 +24,7 @@ public class JmxRuntimeMetricsUtil {
   private static final String INSTRUMENTATION_VERSION =
       EmbeddedInstrumentationProperties.findVersion(INSTRUMENTATION_NAME);
 
-  static Meter getMeter(OpenTelemetry openTelemetry) {
+  public static Meter getMeter(OpenTelemetry openTelemetry) {
     MeterBuilder meterBuilder = openTelemetry.meterBuilder(INSTRUMENTATION_NAME);
     if (INSTRUMENTATION_VERSION != null) {
       meterBuilder.setInstrumentationVersion(INSTRUMENTATION_VERSION);